# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch master
# Your branch is up-to-date with 'origin/master'.
#
# All conflicts fixed but you are still merging.
#
# Changes to be committed:
#	modified:   CHANGELOG.md
#	modified:   TODO.md
#	modified:   docs/commandHelp.rst
#	modified:   docs/examples.rst
#	modified:   docs/faq.rst
#	new file:   docs/screenshots/print_highlight.png
#	modified:   resources/config/black_on_white.conf
#	modified:   resources/config/metal.conf
#	modified:   resources/config/white_on_black.conf
#	modified:   resources/functions.awk
#	modified:   src/coloring/ConfigKey.java
#	modified:   src/coloring/Pdf.java
#	modified:   src/commandHelp/Command.java
#	modified:   src/commandHelp/CommandList.java
#	modified:   src/samTextViewer/ASCIIGenomeHistory.java
#	modified:   src/samTextViewer/ArgParse.java
#	modified:   src/samTextViewer/GenomicCoordsHistory.java
#	modified:   src/samTextViewer/InteractiveInput.java
#	modified:   src/samTextViewer/Main.java
#	modified:   src/samTextViewer/TrackProcessor.java
#	modified:   src/samTextViewer/Utils.java
#	modified:   src/sortBgzipIndex/MakeTabixIndex.java
#	modified:   src/tracks/FeatureFilter.java
#	new file:   src/tracks/Filter.java
#	modified:   src/tracks/GtfSqlite.java
#	modified:   src/tracks/IntervalFeature.java
#	modified:   src/tracks/TextProfile.java
#	modified:   src/tracks/TextRead.java
#	modified:   src/tracks/Track.java
#	modified:   src/tracks/TrackBookmark.java
#	modified:   src/tracks/TrackIntervalFeature.java
#	modified:   src/tracks/TrackPileup.java
#	modified:   src/tracks/TrackReads.java
#	modified:   src/tracks/TrackSeqRegex.java
#	modified:   src/tracks/TrackSet.java
#	modified:   src/tracks/TrackWiggles.java
#	new file:   src/utils/MakeBedgraph.java
#	new file:   test/bashTestFunctions.sh
#	modified:   test/coloring/PdfTest.java
#	modified:   test/commandHelp/CommandListTest.java
#	modified:   test/integration_test.sh
#	modified:   test/samTextViewer/ASCIIGenomeHistoryTest.java
#	modified:   test/samTextViewer/UtilsTest.java
#	modified:   test/sortBgzipIndex/MakeTabixFileTest.java
#	modified:   test/tracks/TextProfileTest.java
#	modified:   test/tracks/TrackIntervalFeatureTest.java
#	modified:   test/tracks/TrackPileupTest.java
#	modified:   test/tracks/TrackReadsTest.java
#	modified:   test/tracks/TrackSetTest.java
#	modified:   test/tracks/TrackTest.java
#	modified:   test/tracks/TrackWigglesTest.java
#	new file:   test/utils/MakeBedgraphTest.java
#	modified:   test_data/asciigenome.yaml
#	new file:   test_data/ear045.oxBS.actb.depth
#	new file:   test_data/empty.bedGraph
#	new file:   test_data/empty2.bedGraph
#	modified:   test_data/invalid-1.bedgraph
#	modified:   test_data/invalid.vcf
#	new file:   test_data/norecords.sam
#	new file:   test_data/norecords.vcf
#	new file:   test_data/norecords_nodict.vcf
#	new file:   test_data/ovl.gff
#	modified:   test_data/variant_reads.sam
#

+++ test/tracks/TextProfileTest.java
@@ -16,40 +16,40 @@ public class TextProfileTest {
 	
 	@Test
 	public void canSetMinMax(){
-		List<Double> yValues= new ArrayList<Double>();
-		yValues.add((double)1);
-		yValues.add((double)2);
-		yValues.add((double)3);
-		yValues.add((double)4);
+		List<Float> yValues= new ArrayList<Float>();
+		yValues.add((float)1);
+		yValues.add((float)2);
+		yValues.add((float)3);
+		yValues.add((float)4);
 		
-		TextProfile tp= new TextProfile(yValues, 10, Double.NaN, Double.NaN);
+		TextProfile tp= new TextProfile(yValues, 10, Float.NaN, Float.NaN);
 		assertEquals(1, tp.getYMinLimit(), 0.001);
 		assertEquals(4, tp.getYMaxLimit(), 0.001);
 		
-		tp= new TextProfile(yValues, 10, -1.0, Double.NaN);
+		tp= new TextProfile(yValues, 10, (float)-1.0, Float.NaN);
 		assertEquals(-1, tp.getYMinLimit(), 0.001);
 		assertEquals(4, tp.getYMaxLimit(), 0.001);
 		
-		tp= new TextProfile(yValues, 10, Double.NaN, 10.0);
+		tp= new TextProfile(yValues, 10, Float.NaN, (float)10.0);
 		assertEquals(1, tp.getYMinLimit(), 0.001);
 		assertEquals(10, tp.getYMaxLimit(), 0.001);
 	}
 	
 	@Test
 	public void canGraphProfile(){
-		List<Double> yValues= new ArrayList<Double>();
-		yValues.add((double)1);
-		yValues.add((double)2);
-		yValues.add((double)3);
-		yValues.add((double)4);
-		yValues.add((double)5);
-		yValues.add((double)6);
-		yValues.add((double)7);
-		yValues.add((double)8);
-		yValues.add((double)9);
-		yValues.add((double)10);
+		List<Float> yValues= new ArrayList<Float>();
+		yValues.add((float)1);
+		yValues.add((float)2);
+		yValues.add((float)3);
+		yValues.add((float)4);
+		yValues.add((float)5);
+		yValues.add((float)6);
+		yValues.add((float)7);
+		yValues.add((float)8);
+		yValues.add((float)9);
+		yValues.add((float)10);
 		
-		TextProfile tp= new TextProfile(yValues, 5, 0.0, Double.NaN);
+		TextProfile tp= new TextProfile(yValues, 5, (float)0.0, Float.NaN);
 		for(List<String> x : tp.getProfile()){
 			System.out.println(x);
 		}
@@ -61,13 +61,13 @@ public class TextProfileTest {
 	@Test
 	public void test() {
 		
-		List<Double> yValues= new ArrayList<Double>();
-		yValues.add((double)0);
-		yValues.add((double)0);
-		yValues.add((double)0);
-		yValues.add((double)0);
+		List<Float> yValues= new ArrayList<Float>();
+		yValues.add((float)0);
+		yValues.add((float)0);
+		yValues.add((float)0);
+		yValues.add((float)0);
 		
-		TextProfile tp= new TextProfile(yValues, 10, Double.NaN, Double.NaN);
+		TextProfile tp= new TextProfile(yValues, 10, Float.NaN, Float.NaN);
 		//System.out.println(tp.getProfile().size());
 		//System.out.println(tp.getProfile());
 	}
